breakout into posthog module
