+++ apps/desktop/src-tauri/src/lib.rs
@@ -16,6 +16,7 @@ mod hotkeys;
 mod notifications;
 mod permissions;
 mod platform;
+mod posthog;
 mod presets;
 mod recording;
 mod recording_settings;
@@ -74,7 +75,6 @@ use std::{
     process::Command,
     str::FromStr,
     sync::Arc,
-    time::Duration,
 };
 use tauri::{AppHandle, Manager, State, Window, WindowEvent, ipc::Channel};
 use tauri_plugin_deep_link::DeepLinkExt;
@@ -1945,14 +1945,7 @@ pub async fn run(recording_logging_handle: LoggingHandle) {
         })
         .ok();
 
-    if let Some(env) = option_env!("VITE_POSTHOG_KEY") {
-        tokio::spawn(async move {
-            posthog_rs::init_global(env)
-                .await
-                .map_err(|err| error!("Error initializing PostHog: {err}"))
-                .ok();
-        });
-    }
+    posthog::init();
 
     let tauri_context = tauri::generate_context!();
 
@@ -2771,45 +2764,4 @@ fn open_project_from_path(path: &Path, app: AppHandle) -> Result<(), String> {
     }
 
     Ok(())
-}
-
-#[derive(Debug)]
-pub enum PostHogEvent {
-    MultpartUploadComplete { duration: Duration },
-    MultpartUploadFailed { duration: Duration, error: String },
-}
-
-impl From<PostHogEvent> for posthog_rs::Event {
-    fn from(event: PostHogEvent) -> Self {
-        match event {
-            PostHogEvent::MultpartUploadComplete { duration } => {
-                let mut e = posthog_rs::Event::new_anon("multipart_upload_complete");
-                e.insert_prop("duration", duration.as_secs())
-                    .map_err(|err| error!("Error adding PostHog property: {err:?}"))
-                    .ok();
-                e
-            }
-            PostHogEvent::MultpartUploadFailed { duration, error } => {
-                let mut e = posthog_rs::Event::new_anon("multipart_upload_failed");
-                e.insert_prop("duration", duration.as_secs())
-                    .map_err(|err| error!("Error adding PostHog property: {err:?}"))
-                    .ok();
-                e.insert_prop("error", error)
-                    .map_err(|err| error!("Error adding PostHog property: {err:?}"))
-                    .ok();
-                e
-            }
-        }
-    }
-}
-
-pub fn async_capture_event(event: PostHogEvent) {
-    if option_env!("VITE_POSTHOG_KEY").is_some() {
-        tokio::spawn(async move {
-            posthog_rs::capture(event.into())
-                .await
-                .map_err(|err| error!("Error sending event to PostHog: {err:?}"))
-                .ok();
-        });
-    }
 }
